requirements.txt
tensorflow==2.10.0
numpy==1.23.4
pillow==9.2.0
sqlite3
imagehash